restrict guest account privileges

+++ client/src/pages/Dashboard/sub-pages/AccountPage/components/DeleteAccount/DeleteAccountButton.js
@@ -1,4 +1,5 @@
 import ThemedButton from '../../../../../../TopLevelComponents/ThemedButton'
+import useSnackbars from "../../../../../../customHooks/useSnackbars"
 import { DeleteUser} from "../../../../../../Services/UserService"
 import React from 'react'
 
@@ -7,7 +8,13 @@ const DeleteAccountButton = (props) => {
     const { userId, history, theme, fromApp } = props
     const { setAuth } = fromApp
 
+    const { disabledForGuestSnackbar } = useSnackbars()
+
     const deleteAccount = async (e) => {
+        if (userId === 20) {
+            disabledForGuestSnackbar()
+            return
+        }
         e.preventDefault()
         await DeleteUser({userId})
         localStorage.clear()
